Bank queue
Challenge
The bank you work at always has trouble organizing customer service queues.

After a meeting with management, it was decided that customers upon arrival at the agency will receive a numeric password on their cell phone via SMS and that the order of the queue will be given not by the order of arrival, but by the number received via SMS. The order of attendance will be decreasing: those who received a larger number must be attended first.

Then, given the order of arrival of customers, reorder the queue according to the number received via SMS, and tell how many customers did not need to change places in this reordering.

Input
The first line contains an integer N, indicating the number of test cases to follow.

Each test case starts with an integer M (1 ≤ M ≤ 1000), indicating the number of customers. Then there will be M distinct integers Pi (1 ≤ Pi ≤ 1000), where the i-th integer indicates the number received via sms from the i-th customer.

The integers above are given in order of arrival, that is, the first integer concerns the first customer to arrive in the queue, the second integer concerns the second customer, and so on.

Output
For each test case, print a line, containing an integer, indicating the number of customers who did not need to change places even after the queue was reordered.

 
Input Example Output Example
3
3
100 80 90
4
100 120 30 50
4
100 90 30 25

1
0
4