Camping Gymkhana
Camping Gymkhana
Challenge
During the December holidays, several schools organize themselves and take their students to a holiday camp for a week. In these camps, students are divided into collective huts by gender and age, always with an adult who, in addition to sleeping with the group in the hut, is also responsible for creating and executing various activities, such as games, excursions, Night Gymkhana, etc. .

On the first day a contest was held in which the activity consisted of grouping students in a circle (organized in a counterclockwise direction) from which they would be removed one by one until there was only one student left, who would be the winner.

The moment each student enters the circle, he / she receives a small chip that contains a value from 1 to 500. After the circle is formed, the number is counted, starting with the student who is next to the first one who entered the circle, the number corresponding to the record the first holds. The student where the counted number falls, must be removed from the group, and the counting starts again according to the student record that has just been eliminated. To make it more interesting, when the value on the card is even, the counting is done clockwise and when the value on the card is odd, the counting is done counterclockwise.

Develop a program so that at the next event, the person responsible for the game will know in advance which child will be the winner of each group, based on the information provided.

Input
The entry contains several test cases. The first line of each test case contains an integer N (1 ≤ N ≤ 100), indicating the number of students who will be part of each circle. Then, the N lines of each test case will contain two pieces of information, the Name and the Value (1 ≤ Value ≤ 500) that appears on each student's file, separated by a space, in the order of entry in the formation of the initial circle.

NOTE: The name of each student must not exceed 30 characters and contains only uppercase and lowercase letters, without accents, and the character "_". The end of the entry is indicated by the number zero.

Output
For each test case, the message Winner (a): xxxxxx must be displayed, with a space after the ":" sign indicating the student in the group who won the game.

 
Input Example Output Example
3
Fernanda 7
Fernando 9
Gustavo 11
5
Mary 7
Peter 9
Joao_Vitor 5
Isabel 12
Laura 8
3
Mary 4
Peter 3
John 2
0

Winner: Fernanda
Winner: Pedro
Winner: Pedro