Password validator pattern
Challenge
Pedro and Fernando are the developers in a stratup and will develop the new registration system, and asked for your help. Your task is to make the code that validates the passwords that are registered, for that you must pay attention to the following requirements:

The password must contain at least one uppercase letter, one lowercase letter and a number;
It cannot have any punctuation, accent or space characters;
In addition, the password can be 6 to 32 characters long.
Input
The entry contains several test cases and ends with the end of the file. Each line has an S string, corresponding to the password that is entered by the user at the time of registration.

Output
The output contains a line, which can be “Password valid.”, If the password has each item of the previously requested requirements, or “Password invalid.”, If one or more requirements are not met.

 
Input Example Output Example
Digital Innovation One

AbcdEfgh99

DigitalInnovationOne123

Digital Innovation One 123

Aass9

Aassd9

Invalid password.

Password valid.

Password valid.

Invalid password.

Invalid password.

Password valid.